Clippy issues fixed

+++ crates/tako/src/internal/worker/test_util.rs
@@ -102,7 +102,7 @@ impl ResourceQueueBuilder {
         self.queue
             .try_start_tasks(&self.task_map, None)
             .into_iter()
-            .map(|(t, a, _)| (t.as_num() as u64, a))
+            .map(|(t, a, _)| (t.as_num(), a))
             .collect()
     }
 
@@ -110,7 +110,7 @@ impl ResourceQueueBuilder {
         self.queue
             .try_start_tasks(&self.task_map, Some(duration))
             .into_iter()
-            .map(|(t, a, _)| (t.as_num() as u64, a))
+            .map(|(t, a, _)| (t.as_num(), a))
             .collect()
     }
 }
